+++ frontend/src/app/auth/http-interceptor.service.ts
@@ -2,6 +2,7 @@ import { HttpErrorResponse, HttpInterceptorFn } from '@angular/common/http';
 import { inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { catchError, throwError } from 'rxjs';
+import Swal from 'sweetalert2';
 
 export const meuhttpInterceptor: HttpInterceptorFn = (request, next) => {
 
@@ -22,11 +23,14 @@ export const meuhttpInterceptor: HttpInterceptorFn = (request, next) => {
         console.log('entrou aqui 2');
         
         if (err.status === 401) {
-          alert('401 - tratar aqui');
-          router.navigate(['/login']);
-        } else
-        if (err.status === 403) {
-          alert('403 - tratar aqui');
+          if (router.url.includes('/login')) {
+            Swal.fire({ title: 'Usuário ou senha incorretos!', icon: 'error', confirmButtonText: 'Ok' });
+          } else {
+            Swal.fire({ title: 'Sessão expirada!', icon: 'error', confirmButtonText: 'Ok' });
+            router.navigate(['/login']);
+          }
+        } else if (err.status === 403) {
+          Swal.fire({ title: 'Acesso negado!', icon: 'error', confirmButtonText: 'Ok' });
           router.navigate(['/login']);
         } else {
           console.error('HTTP error:', err);
