+++ frontend/src/app/components/layout/login/login.component.ts
@@ -3,8 +3,9 @@ import { FormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
 import { MdbFormsModule } from 'mdb-angular-ui-kit/forms';
 import { MdbRippleModule } from 'mdb-angular-ui-kit/ripple';
-import { LoginService } from '../../../auth/login.service';
-import { Login } from '../../../auth/login';
+import Swal from 'sweetalert2';
+import { UsuariosService } from '../../../services/usuarios.service';
+import { Login } from '../../../models/login';
 
 @Component({
   selector: 'app-login',
@@ -16,26 +17,23 @@ export class LoginComponent {
   login: Login = new Login;
 
   router = inject(Router);
-  loginService = inject(LoginService);
+  usuariosService = inject(UsuariosService);
 
   logar(){
-
-    this.loginService.logar(this.login).subscribe({
-
+    this.usuariosService.login(this.login).subscribe({
       next: token => {
-
         if(token){
-          this.loginService.addToken(token);
-        }else{
-          alert('Email ou senha incorretos!')
+          this.usuariosService.addToken(token);
+          this.router.navigate(['/admin/turmas']);
         }
-
       },
       error: erro => {
-        alert('ERRO');
-
+        if (erro.status === 401) {
+          Swal.fire({ title: 'Usuário ou senha incorretos!', icon: 'error', confirmButtonText: 'Ok' });
+        } else {
+          Swal.fire({ title: 'Ocorreu um erro!', icon: 'error', confirmButtonText: 'Ok' });
+        }
       }
-
     });
 
   }
