+++ frontend/src/app/models/login.ts
@@ -1,4 +1,3 @@
-
 export class Login {
   email!: string;
   senha!: string;
